+++ js/data.js
@@ -1,0 +1,60 @@
+const firstNames = [
+  'Айрат',
+  'Борис',
+  'Алёна',
+  'Всеволод',
+  'Бенджамин',
+  'Гарри',
+  'Дамир',
+  'Октябрина',
+  'Илья',
+  'Нина',
+  'Ким',
+  'Августа',
+  'Айсун',
+  'Николь',
+];
+
+const lastNames = [
+  'Смит',
+  'Тейлор',
+  'Джексон',
+  'Харрис',
+  'Мартинес',
+  'Ли',
+  'Уокер',
+  'Эрнандес',
+  'Райт',
+  'Грин',
+  'Филлипс',
+  'Кэмпбелл',
+  'Паркер',
+  'Эванс',
+]
+
+const descriptions = [
+  'Ножки крестиком , лапки бантиком',
+  'Рыженькие кошки - пушистенькие ножки.',
+  'Я на работе)',
+  'Плюша',
+  'Вот где грация!',
+  'Ой, какой безмятежный сон у котишки',
+];
+
+const messages = [
+  'Всё отлично!',
+  'В целом всё неплохо. Но не всё.',
+  'Когда вы делаете фотографию, хорошо бы убирать палец из кадра. В конце концов это просто непрофессионально.',
+  'Моя бабушка случайно чихнула с фотоаппаратом в руках и у неё получилась фотография лучше.',
+  'Я поскользнулся на банановой кожуре и уронил фотоаппарат на кота и у меня получилась фотография лучше.',
+  'Лица у людей на фотке перекошены, как будто их избивают. Как можно было поймать такой неудачный момент?!',
+];
+
+const MIN_LIKES = 15;
+const MAX_LIKES = 200;
+const COUNT = 25;
+const MAX_AVATAR = 6;
+const ids = [];
+
+
+export {firstNames, lastNames, descriptions, messages, MIN_LIKES, MAX_LIKES, COUNT, MAX_AVATAR, ids};
