+++ js/main.js
@@ -1,85 +1,5 @@
-// Функция, возвращающая случайное целое число из переданного диапазона включительно.
-// https://stackoverflow.com/questions/1527803/generating-random-whole-numbers-in-javascript-in-a-specific-range
-
-const getRandomNumber = function(min, max) {
-
-  if(min<max) {
-    min = Math.ceil(min);
-    max = Math.floor(max);
-    return Math.floor(Math.random() * (max - min + 1)) + min;
-  }
-  throw new Error('Значение ' + min + ' не может быть больше, чем значение ' + max + '!')
-}
-
-// Функция для проверки максимальной длины строки.
-// Тут не сложно, сама написала функцию
-
-const isStringTooLong = function(stringToCheck, maxLength) {
-  const stringToCheckLength = stringToCheck.length;
-
-  return stringToCheckLength > maxLength;
-};
-
-isStringTooLong('casjdajsdjajd ajdjasd asjdjasdjas asdahsdhd', 140);
-
-const firstNames = [
-  'Айрат',
-  'Борис',
-  'Алёна',
-  'Всеволод',
-  'Бенджамин',
-  'Гарри',
-  'Дамир',
-  'Октябрина',
-  'Илья',
-  'Нина',
-  'Ким',
-  'Августа',
-  'Айсун',
-  'Николь',
-];
-
-const lastNames = [
-  'Смит',
-  'Тейлор',
-  'Джексон',
-  'Харрис',
-  'Мартинес',
-  'Ли',
-  'Уокер',
-  'Эрнандес',
-  'Райт',
-  'Грин',
-  'Филлипс',
-  'Кэмпбелл',
-  'Паркер',
-  'Эванс',
-]
-
-const descriptions = [
-  'Ножки крестиком , лапки бантиком',
-  'Рыженькие кошки - пушистенькие ножки.',
-  'Я на работе)',
-  'Плюша',
-  'Вот где грация!',
-  'Ой, какой безмятежный сон у котишки',
-];
-
-const messages = [
-  'Всё отлично!',
-  'В целом всё неплохо. Но не всё.',
-  'Когда вы делаете фотографию, хорошо бы убирать палец из кадра. В конце концов это просто непрофессионально.',
-  'Моя бабушка случайно чихнула с фотоаппаратом в руках и у неё получилась фотография лучше.',
-  'Я поскользнулся на банановой кожуре и уронил фотоаппарат на кота и у меня получилась фотография лучше.',
-  'Лица у людей на фотке перекошены, как будто их избивают. Как можно было поймать такой неудачный момент?!',
-];
-
-const MIN_LIKES = 15;
-const MAX_LIKES = 200;
-const COUNT = 25;
-const MAX_AVATAR = 6;
-const ids = [];
-
+import {firstNames, lastNames, descriptions, messages, MIN_LIKES, MAX_LIKES, COUNT, MAX_AVATAR, ids} from './data.js';
+import {getRandomNumber} from './util.js';
 
 const getRandomElement = (array) => {
   return array[getRandomNumber(0, array.length-1)];
@@ -115,7 +35,7 @@ const generateComments = (photoId) => {
   return result;
 }
 
-const generatePhotoDescription = ids.map(id => (
+const photoDescriptionMocks = ids.map(id => (
   {
     id: id,
     url: `photos/${id}.jpg`,
@@ -125,4 +45,4 @@ const generatePhotoDescription = ids.map(id => (
   }),
 )
 
-generatePhotoDescription();
+export {photoDescriptionMocks};
