+++ js/script.js
@@ -1,0 +1,3 @@
+import {photoDescriptionMocks} from './main.js';
+
+photoDescriptionMocks;
