+++ js/util.js
@@ -1,0 +1,19 @@
+const getRandomNumber = function(min, max) {
+
+  if(min<max) {
+    min = Math.ceil(min);
+    max = Math.floor(max);
+    return Math.floor(Math.random() * (max - min + 1)) + min;
+  }
+  throw new Error('Значение ' + min + ' не может быть больше, чем значение ' + max + '!')
+}
+
+const isStringTooLong = function(stringToCheck, maxLength) {
+  const stringToCheckLength = stringToCheck.length;
+
+  return stringToCheckLength > maxLength;
+};
+
+isStringTooLong('casjdajsdjajd ajdjasd asjdjasdjas asdahsdhd', 140);
+
+export {getRandomNumber};
